requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0
pyspellchecker>=0.8.0